+++ changelog.md
@@ -12,6 +12,7 @@ Codion Change Log
 - NumberField.Builder.groupingUsed() renamed grouping().
 - Remaining usages of SwingConstants.LEFT and RIGHT replaced with LEADING and TRAILING.
 - LabelBuilder.HORIZONTAL_ALIGNMENT removed.
+- NumberDocument, value included in range validation exception message.
 ### is.codion.framework.domain
 - KeyGenerator.Identity removed.
 - AttributeDefinition.Builder.valueRange(), minimumValue() and maximumValue() renamed range(), minimum() and maximum().